Early Development Build: 15/06/2018
Ability 2 Functionality Complete.
Animations + FX needed.

+++ Assets/Scripts/Enemies/Enemy.cs
@@ -20,7 +20,7 @@ public class Enemy : MonoBehaviour
         Death();
     }
 
-   public virtual void TakeDamage(int amount)
+   public virtual void TakeDamage(float amount)
     {
         health -= amount; // Cast floats to int
     }
